Update Authentication Administrator role
The documentation indicates:

Authentication administrators can require users who are non-administrators or assigned to some roles to re-register against existing non-password credentials (for example, MFA or FIDO), and can also revoke remember MFA on the device, which prompts for MFA on the next sign-in.

Based on test, a user with Authentication administrator role cannot require another user to re-register MFA. However, they can reset the user password (even for admin accounts).
Test:
Added Authentication administrator role to a user account.
MFA was Enabled for test user. 
Option to Require re-register MFA is greyed out (this applies for admin and regular user accounts)
Same as the option to access MFA settings:

This gap between documentation and reality was reported by a customer, and based on that we tested it. 

Thank you